requirements.txt
# Dev environment
pip
autopep8

# App
flask
boto3
openai
python-dotenv